Added property for content type. creating folder in isoloated storage for every msisdn to store file attachments for particular msisdn

+++ windows-client/Model/Attachment.cs
@@ -12,6 +12,7 @@ using System.Windows.Media.Imaging;
 using System.IO;
 using System.Windows.Resources;
 using System.IO.IsolatedStorage;
+using System.ComponentModel;
 
 namespace windows_client.Model
 {
@@ -19,6 +20,7 @@ namespace windows_client.Model
     {
         private string _fileKey;
         private string _fileName;
+        private string _contentType;
         private byte[] _thumbnail;
 
         public string FileKey
@@ -51,6 +53,21 @@ namespace windows_client.Model
             }
         }
 
+        public string ContentType
+        {
+            get
+            {
+                return _contentType;
+            }
+            set
+            {
+                if (_contentType != value)
+                {
+                    _contentType = value;
+                }
+            }
+        }
+
         public byte[] Thumbnail
         {
             get
@@ -66,51 +83,91 @@ namespace windows_client.Model
             }
         }
 
-        public Attachment(string fileName, string fileKey, byte[] thumbnailBytes)
+        //newly received messages
+        public Attachment(string fileName, string fileKey, byte[] thumbnailBytes, string contentType)
         {
             this.FileName = fileName;
             this.FileKey = fileKey;
             this.Thumbnail = thumbnailBytes;
-
-
-            //run a bg to download image and store to isolated storage
+            this.ContentType = contentType;
         }
 
-        public void storeThumbnailInIsolatedStorage(long msgId)
+        public Attachment(string fileName, byte[] thumbnailBytes)
         {
-            string filePath = HikeConstants.FILE_TRANSFER + "/" + Convert.ToString(msgId);
-            using (IsolatedStorageFile myIsolatedStorage = IsolatedStorageFile.GetUserStoreForApplication())
-            {
-                if (myIsolatedStorage.FileExists(filePath))
-                {
-                    myIsolatedStorage.DeleteFile(filePath);
-                }
+            this.FileName = fileName;
+            this.Thumbnail = thumbnailBytes;
+        }
 
-                using (IsolatedStorageFileStream fileStream = new IsolatedStorageFileStream(filePath, FileMode.Create, myIsolatedStorage))
+
+        public static void storeFileInIsolatedStorage(string filePath, byte[] imagebytes)
+        {
+            string fileDirectory = filePath.Substring(0, filePath.LastIndexOf("/"));
+            if (imagebytes != null)
+            {
+                using (IsolatedStorageFile myIsolatedStorage = IsolatedStorageFile.GetUserStoreForApplication())
                 {
-                    using (BinaryWriter writer = new BinaryWriter(fileStream))
+                    if (!myIsolatedStorage.DirectoryExists(fileDirectory))
                     {
-                        writer.Write(_thumbnail, 0, _thumbnail.Length);
+                        myIsolatedStorage.CreateDirectory(fileDirectory);
+                    }
+
+                    if (myIsolatedStorage.FileExists(filePath))
+                    {
+                        myIsolatedStorage.DeleteFile(filePath);
+                    }
+
+                    using (IsolatedStorageFileStream fileStream = new IsolatedStorageFileStream(filePath, FileMode.Create, myIsolatedStorage))
+                    {
+                        using (BinaryWriter writer = new BinaryWriter(fileStream))
+                        {
+                            writer.Write(imagebytes, 0, imagebytes.Length);
+                        }
                     }
                 }
             }
         }
 
-        public Attachment(long msgId)
-        {
-            string filePath = HikeConstants.FILE_TRANSFER + "/" + Convert.ToString(msgId);
-
+        public static string[] getAttachmentFiles(string msisdn)
+        { 
+            string filePath = HikeConstants.FILE_TRANSFER_LOCATION + "/" + "Attachments" + "/" + msisdn + "/*";
+            string[] fileNames = null;
             using (IsolatedStorageFile myIsolatedStorage = IsolatedStorageFile.GetUserStoreForApplication())
             {
-                using (IsolatedStorageFileStream fileStream = myIsolatedStorage.OpenFile(filePath, FileMode.Open, FileAccess.Read))
+                if (myIsolatedStorage.FileExists(filePath))
                 {
-                    _thumbnail = new byte[fileStream.Length];
-                    // Read the entire file and then close it
-                    fileStream.Read(_thumbnail, 0, _thumbnail.Length);
-                    fileStream.Close();
+                    fileNames = myIsolatedStorage.GetFileNames(filePath);
+                }
+            }
+            return fileNames;        
+        }
+
+
+        public static void readFileFromIsolatedStorage(string filePath, out byte[] imageBytes)
+        {
+            using (IsolatedStorageFile myIsolatedStorage = IsolatedStorageFile.GetUserStoreForApplication())
+            {
+                if (myIsolatedStorage.FileExists(filePath))
+                {
+                    using (IsolatedStorageFileStream fileStream = myIsolatedStorage.OpenFile(filePath, FileMode.Open, FileAccess.Read))
+                    {
+                        imageBytes = new byte[fileStream.Length];
+                        // Read the entire file and then close it
+                        fileStream.Read(imageBytes, 0, imageBytes.Length);
+                        fileStream.Close();
+                    }
+                }
+                else
+                {
+                    imageBytes = null;
                 }
             }
         }
+
+        //while reading messages from db
+        //public Attachment(long msgId)
+        //{
+        //    readFileFromIsolatedStorage(HikeConstants.FILE_TRANSFER_LOCATION + "/" + Convert.ToString(msgId), out _thumbnail);
+        //}
 
     }
 }
